requirements.txt
Flask~=1.1.2
gunicorn==19.10.0
redis~=3.5.3
Flask-Session==0.3.2
python-dotenv==0.15.0
bcrypt==3.2.0
pyopenssl==19.1.0
cryptography==3.2.1